testing/web-platform/tests/workers/WorkerGlobalScope_importScripts_NetworkErr.htm

<!DOCTYPE html>
<html>
<head>
<title> importScripts() with non-existent script file raises NETWORK_ERR </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>
<script>
    
    var FileName = "./support/ImportScriptsNetworkErr.js";
    var ExpectedResult = "Pass";
    var description = "importScripts(): Throw NETWORK_ERR exception whenever attempt to fetch script is failed.";
    
    var t = async_test("Test Description: " + description);
                  
    var worker = new Worker(FileName);

    worker.onmessage = t.step_func(function(evt)
    {
        assert_equals(evt.data, ExpectedResult);
        t.done();
    });
</script>
</body>
</html>
